+++ apps/user-front/src/tabs/search/components/AutoComplete.tsx
