+++ apps/user-front/src/tabs/search/components/EmptyStoreState.tsx
